LOST COMMITS (due to repo split):
koerberm	
Fixed hanging queries bug when node crashes! 

koerberm	
moved rescheduling of aborted jobs to query_manager; 

koerbem
Using std::numeric_limits<double>::infinity() to indicate a stopped
timer --> avoids trouble with low resolution timers;

koerbem
Fixed puzzling of simple feature collections;

mattig
Implement Serialization and Deserialization for Plots and the Histogram

mattig	
forgot plot.cpp 

mattig	
test for plot serialization 

koerberm	
raised log-level to trace for repeating statistics message; 

koerberm	
Delegating histogramm creation to workers; 

koerberm
Fixed bug when worker queried index-cache with other types than RASTER;
Implemented exact search to cache plots;

mattig	
Configuration: Add getLong method 

mattig	
Cache: Load cache size config as Long to support larger caches than int
max

+++ cache/node/node_config.cpp
@@ -21,11 +21,11 @@ NodeConfig NodeConfig::fromConfiguration() {
 	result.caching_strategy = Configuration::get("nodeserver.cache.strategy");
 	result.local_replacement = Configuration::get("nodeserver.cache.local.replacement", "lru");
 
-	result.raster_size = Configuration::getInt("nodeserver.cache.raster.size");
-	result.point_size = Configuration::getInt("nodeserver.cache.points.size");
-	result.line_size = Configuration::getInt("nodeserver.cache.lines.size");
-	result.polygon_size = Configuration::getInt("nodeserver.cache.polygons.size");
-	result.plot_size = Configuration::getInt("nodeserver.cache.plots.size");
+	result.raster_size = Configuration::getLong("nodeserver.cache.raster.size");
+	result.point_size = Configuration::getLong("nodeserver.cache.points.size");
+	result.line_size = Configuration::getLong("nodeserver.cache.lines.size");
+	result.polygon_size = Configuration::getLong("nodeserver.cache.polygons.size");
+	result.plot_size = Configuration::getLong("nodeserver.cache.plots.size");
 	return result;
 }
